move `utils.DataUnit.DataRate` to `utils.DataRate`
